Arreglado error, que usaba la variable incorrecta para el control de votos.
git-svn-id: file:///home/gallir/meneame/svn/src@661 cbe1a1ec-7179-4919-8e32-4551e8279a00

+++ branches/version2/www/libs/comment.php
@@ -148,7 +148,7 @@ class Comment {
 
 	function print_shake_icons() {
 		global $globals, $current_user;
-		if ( $this->date > time() - $globals['time_enabled_votes'] && ! $this->vote_exists()) {  
+		if ( $current_user->user_karma > $globals['min_karma_for_comment_votes'] && $this->date > time() - $globals['time_enabled_votes'] && ! $this->vote_exists()) {  
 		 	echo '<span id="c-votes-'.$this->id.'">';
 			echo '<a href="javascript:menealo_comment('."$current_user->user_id,$this->id,1".')" title="'._('voto positivo').'"><img src="'.$globals['base_url'].'img/common/vote-up01.png" width="12" height="12" alt="'._('voto positivo').'"/></a>&nbsp;&nbsp;&nbsp;';
 		 	echo '<a href="javascript:menealo_comment('."$current_user->user_id,$this->id,-1".')" title="'._('voto negativo').'"><img src="'.$globals['base_url'].'img/common/vote-down01.png" width="12" height="12" alt="'._('voto negativo').'"/></a>&nbsp;';
